+++ src/Content/phpwasm/MyPHPWebsite.1.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31605.320
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Client", "Client\Client.csproj", "{059EF486-DE6A-4A48-AF61-55EEDEE16815}"
+EndProject
+Project("{13B669BE-BB05-4DDF-9536-439F39A36129}") = "PHPScripts", "PHPScripts\PHPScripts.msbuildproj", "{70E556DA-E4E0-483E-B493-06AF532653ED}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Server", "Server\Server.csproj", "{DFC33DD0-9F3A-4A2A-BAF4-04537E20C75D}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{059EF486-DE6A-4A48-AF61-55EEDEE16815}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{059EF486-DE6A-4A48-AF61-55EEDEE16815}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{059EF486-DE6A-4A48-AF61-55EEDEE16815}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{059EF486-DE6A-4A48-AF61-55EEDEE16815}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70E556DA-E4E0-483E-B493-06AF532653ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70E556DA-E4E0-483E-B493-06AF532653ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70E556DA-E4E0-483E-B493-06AF532653ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70E556DA-E4E0-483E-B493-06AF532653ED}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DFC33DD0-9F3A-4A2A-BAF4-04537E20C75D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DFC33DD0-9F3A-4A2A-BAF4-04537E20C75D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DFC33DD0-9F3A-4A2A-BAF4-04537E20C75D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DFC33DD0-9F3A-4A2A-BAF4-04537E20C75D}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6E295449-F228-4630-BF24-AF174081200F}
+	EndGlobalSection
+EndGlobal
