--- src/Content/phpwasm/PHPScripts/about/me.php
@@ -1,12 +1,0 @@
-<?php
-    require("/headers/defaultHeader.php");
-?>
-
-<h1>Me</h1>
-<p>Something about me...</p>
-<p>This image will not be displayed, when you are in the offline mode</p>
-<img alt="Logo" src="images/logo.png"/>
-
-<?php
-    require("/footers/defaultFooter.php");
-?>
