--- src/Content/phpwasm/PHPScripts/community/developerDanesh.php
@@ -1,2 +1,0 @@
-<h1>Danesh</h1>
-<p>Something about Danesh...</p>
