--- src/Content/phpwasm/PHPScripts/community/developerRichard.php
@@ -1,2 +1,0 @@
-<h1>Richard</h1>
-<p>Something about Richard...</p>
