--- src/Content/phpwasm/PHPScripts/footers/defaultFooter.php
@@ -1,2 +1,0 @@
-</body>
-</html>
