--- src/Content/phpwasm/PHPScripts/headers/defaultHeader.php
@@ -1,9 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-	<title>PHP WEB</title>
-</head>
-<body>
-<?php
-    require("/layouts/defaultLayout.php");
-?>
