--- src/Content/phpwasm/PhpWasm.1.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31019.35
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorApp.Server", "BlazorApp\Server\BlazorApp.Server.csproj", "{973C6AA9-5C6A-4A30-871B-B031E0E282B9}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BlazorApp.Client", "BlazorApp\Client\BlazorApp.Client.csproj", "{9D1A1057-1AC2-4306-8E36-12191AC95505}"
-EndProject
-Project("{13B669BE-BB05-4DDF-9536-439F39A36129}") = "PHPScripts", "PHPScripts\PHPScripts.msbuildproj", "{F0CA7993-44E4-4025-A63B-BFAB0EC8667E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{973C6AA9-5C6A-4A30-871B-B031E0E282B9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{973C6AA9-5C6A-4A30-871B-B031E0E282B9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{973C6AA9-5C6A-4A30-871B-B031E0E282B9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{973C6AA9-5C6A-4A30-871B-B031E0E282B9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9D1A1057-1AC2-4306-8E36-12191AC95505}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9D1A1057-1AC2-4306-8E36-12191AC95505}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9D1A1057-1AC2-4306-8E36-12191AC95505}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9D1A1057-1AC2-4306-8E36-12191AC95505}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F0CA7993-44E4-4025-A63B-BFAB0EC8667E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F0CA7993-44E4-4025-A63B-BFAB0EC8667E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F0CA7993-44E4-4025-A63B-BFAB0EC8667E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F0CA7993-44E4-4025-A63B-BFAB0EC8667E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {7396B75F-9F65-4250-A3B8-350F107FD063}
-	EndGlobalSection
-EndGlobal
